requirements.txt
# OpenAI Agents SDK
git+https://github.com/openai/openai-agents-sdk.git

# Web Interface
gradio>=4.0.0

# Environment Variables
python-dotenv>=1.0.0

# Email Service
sendgrid>=6.11.0

# Data Validation
pydantic>=2.0.0

# Async Support
asyncio
